% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Create dataset descriptions for roxygen.  The workhorse.}
\usage{
describe(data, text, file, outfile, fileEncoding = "", envir = NULL,
  encoding = "unknown", comment.char = "##'", start = "\\\\itemize {",
  end = "}", item = "\\\\item \%s:", indent = 3, sublist = NULL,
  add_tables = NULL, ...)
}
\arguments{
\item{data}{Input data}

\item{text}{Optional text describing dataset (to go in description area).}

\item{file}{Optional location of file with description of data}

\item{outfile}{Optional location of file to write output}

\item{fileEncoding}{File encoding}

\item{envir}{Environment to find data}

\item{encoding}{encoding}

\item{comment.char}{prefix character for roxygen template (default )}

\item{start}{Start of itemized list (default "\\\\itemize \{")}

\item{end}{End of itemized list (default "\}")}

\item{item}{Format for an item in list as regex to pass to sprintf.}

\item{indent}{How many spaces to indent itemized lists (default: 3)}

\item{sublist}{Optional list of columns (factors/characters) that should be sublisted.}

\item{add_tables}{List of columns to add tabular counts in description (or TRUE for all).}

\item{...}{further arguments passed to roxy_format}
}
\value{
Vector of lines for roxygen metadata, or write to file if \code{outfile}.
}
\description{
Create dataset descriptions for roxygen.  The workhorse.
}
\examples{
## run devtools::document() afterward for example
\dontrun{
  roxy_itemize(mtcars, text='Dataset about cars', outfile='temp.R', add_tables=TRUE)
}
}

