% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{roxy_format}
\alias{roxy_format}
\title{Generic formatter for roxy_itemize variable types}
\usage{
roxy_format(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{passed to dispatchers}
}
\description{
Generic formatter for roxy_itemize variable types
}

