% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{roxy_tabular}
\alias{roxy_tabular}
\title{Similar to http://r-pkgs.had.co.nz/man.html to make tabular roxy output.
Roxygen tabulate}
\usage{
roxy_tabular(df, ...)
}
\arguments{
\item{df}{data.frame}

\item{...}{arguments passed to format}
}
\value{
character vector of lines for table
}
\description{
Similar to http://r-pkgs.had.co.nz/man.html to make tabular roxy output.
Roxygen tabulate
}

